+++ components/sensors/libs/co2_sensor.h
@@ -37,13 +37,13 @@ esp_err_t co2_init(co2_sensor_t *dev, i2c_port_t port, uint8_t addr, int8_t sda_
  * @param co2 pointer to co2 variable
  * @return ESP_OK to indicate success
  */
-esp_err_t read_co2(co2_sensor_t *dev, unsigned int * co2);
+esp_err_t read_co2(co2_sensor_t *dev, float * co2);
 
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif /* PH_SENSOR_H */
+#endif /* CO@_SENSOR_H */
